Moved IKanjiNoteService to correct folder.
